+++ src/lib/utils.ts
@@ -36,6 +36,7 @@ import {
 import {
 	APIFY_CLIENT_DEFAULT_HEADERS,
 	AUTH_FILE_PATH,
+	CommandExitCodes,
 	DEFAULT_LOCAL_STORAGE_DIR,
 	GLOBAL_CONFIGS_FOLDER,
 	INPUT_FILE_REG_EXP,
@@ -118,9 +119,9 @@ export async function getLoggedClientOrThrow() {
 	const loggedClient = await getLoggedClient();
 
 	if (!loggedClient) {
+		process.exitCode = CommandExitCodes.MissingAuth;
 		throw new Error('You are not logged in with your Apify account. Call "apify login" to fix that.');
 	}
-
 	return loggedClient;
 }